// front/src/components/card/types.ts
export interface RecordCard {
  record: {
    subject: {
      problem: [];
      situation: string;
      comments: string;
    };
    object: {
      problem: [];
      situation: string;
      comments: string;
    };
    assessment: {
      problem: [];
      situation: string;
      comments: string;
    };
    plan: {
      problem: [];
      situation: string;
      comments: string;
    };
  };
}